requirements.txt
langchain
langchain_community
langchain_core
pandas
pyarrow
pyprojroot
python-dotenv
PyYAML
SQLAlchemy
openai
pip-system-certs

sqllite3
ipython-sql

azure-identity

langchain-openai
langgraph

pydantic_ai

aiosqlite

chainlit

azure-ai-inference
networkx